re organised and refactored small part of _Resolve() func to only use one for loop to get active context
