+++ src/AyonUsdResolver/cache/assetIdentDef.cpp
