+++ src/AyonUsdResolver/cache/assetIdentDef.h
