+++ src/AyonUsdResolver/pluginData/module.cpp
@@ -1,8 +1,6 @@
-#include "pxr/pxr.h"
 #include "pxr/base/tf/pyModule.h"
 
-TF_WRAP_MODULE
-{
+TF_WRAP_MODULE {
     TF_WRAP(Resolver);
     TF_WRAP(ResolverContext);
     TF_WRAP(ResolverTokens);
